修改unit为uint64_t

+++ alc/proto.h
@@ -1,5 +1,5 @@
 // astnodes.c
-typedef unsigned int uint;
+typedef uint64_t uint;
 
 ASTnode *mkastnode(int op, ASTnode *left, ASTnode *mid, ASTnode *right);
 ASTnode *mkastleaf(int op, Type *type, bool rvalue, Sym *sym, uint64_t uintval);
